Added ability to render X-based sprites (plants). Some improvements to the FrameBuffer implementation and API. Added ability to traverse a ray through the world, which enables functionality like breaking and placing blocks.

+++ src/graphics/FrameBuffer.cpp
@@ -31,6 +31,6 @@ namespace zore {
 	}
 
 	uint FrameBuffer::GetNumTextureAttachments() {
-		return attachments.size();
+		return static_cast<uint>(attachments.size());
 	}
 }
